Final State

+++ Assets/Shaders/Ocean.shader
@@ -73,9 +73,8 @@ Shader "Custom/Ocean"
             v.vertex.y += waveX * waveZ * _WaveAmplitude;
             
             //Ripple Effect
-
-            float offsetvert = ((v.vertex.x *  v.vertex.x)+(v.vertex.z *  v.vertex.z))+ ((v.vertex.x * _OffSetX) + (v.vertex.z * _OffSetZ)) ;
-            float value = sin(_Time.w * _RippleSpeed  + offsetvert * _RipplePeriod );
+            half offsetvert = ((v.vertex.x * v.vertex.x) + (v.vertex.z * v.vertex.z))+ ((v.vertex.x * _OffSetX) + (v.vertex.z * _OffSetZ));
+            float value = sin(_Time.w * _RippleSpeed + offsetvert * _RipplePeriod   );
             float3 worldPos = mul(unity_ObjectToWorld, v.vertex ).xyz;
 
             if( sqrt(pow(worldPos.x-_ImpactX,2)+pow(worldPos.z-_ImpactZ,2)) < _Distance){
